+++ components/button/NavButton.tsx
@@ -9,13 +9,24 @@ type NavButtonProps = {
   text: string;
   size: string;
   link: string;
+  newTab?: boolean;
 };
 
-export default function NavButton({ text, size, link }: NavButtonProps) {
+export default function NavButton({
+  text,
+  size,
+  link,
+  newTab,
+}: NavButtonProps) {
   const { isVisible, targetRef } = useFloatingEffect<HTMLAnchorElement>();
 
   return (
-    <Link className={styles["Link"]} href={link} ref={targetRef}>
+    <Link
+      className={styles["Link"]}
+      href={link}
+      ref={targetRef}
+      target={`${newTab ? "_blank" : ""}`}
+    >
       <div
         className={`${styles["button-wrapper"]} ${
           styles[getButtonSizeClass(size) as string]
