+++ components/button/button.module.css
@@ -7,6 +7,7 @@
   z-index: 1;
   transition: all 0.4s ease;
   transform: translateY(2rem);
+  background-color: var(--primary-color);
 
   border: 1px solid var(--color-light-gray);
   border-radius: 20px;
