+++ components/main-page/about-post-section/dev-posts/dev-posts.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from "react";
 
 import styles from "./dev-posts.module.scss";
-import NavButton from "@/components/button/nav-button";
+import NavButton from "@/components/button/NavButton";
 import Carousel from "@/components/carousel/carousel";
 import { sortRecentDevPosts } from "@/utils/sortPosts";
 import { fetchPostList } from "@/utils/api";
