+++ components/main-page/about-post-section/recent-posts/recent-posts.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from "react";
 
 import styles from "./recent-posts.module.scss";
-import NavButton from "@/components/button/nav-button";
+import NavButton from "@/components/button/NavButton";
 import Carousel from "@/components/carousel/carousel";
 import { sortRecentPosts } from "@/utils/sortPosts";
 import { fetchPostList } from "@/utils/api";
